Fixed AddEntityPacket

+++ src/network/protocol/packet/AddEntityPacket.php
@@ -41,7 +41,7 @@ class AddEntityPacket extends RakNetDataPacket{
 	public function encode(){
 		$this->reset();
 		$this->putInt($this->eid);
-		$this->putInt($this->type);
+		$this->putByte($this->type);
 		$this->putFloat($this->x);
 		$this->putFloat($this->y);
 		$this->putFloat($this->z);
